changed location of files for better overview
